create: create inner drawer
