+++ jin/screens/Chat.js
@@ -3,11 +3,8 @@ import { GiftedChat } from 'react-native-gifted-chat'
 import firestore from '@react-native-firebase/firestore'
 import auth from '@react-native-firebase/auth'
 import { NativeBaseProvider, Modal, Button } from 'native-base';
-import { createDrawerNavigator } from '@react-navigation/drawer'
 
-const Drawer = createDrawerNavigator();
-
-function Chat({ navigation, route }) {
+export default function Chat({ navigation, route }) {
   const { thread } = route.params;
   const user = auth().currentUser;
   const [messages, setMessages] = useState([]);
@@ -142,38 +139,4 @@ function Chat({ navigation, route }) {
       </Modal>
     </NativeBaseProvider>
   );
-}
-
-export default function ChatDrawer({ route }) {
-  return (
-    <Drawer.Navigator
-      screenOptions={{
-        drawerPosition: 'right'
-      }}
-    >
-      <Drawer.Screen
-        name="메시지 화면"
-        component={Chat}
-        initialParams={route.params}
-      // options={({ route }) => ({
-      //   headerTitle: () => (
-      //     <Text style={styles.headerTitle} bold>{route.params.thread.name}</Text>
-      //   ),
-      //   headerBackTitleVisible: false,
-      //   headerStyle: {
-      //     backgroundColor: '#BF2A52',
-      //   },
-      //   headerTintColor: '#fff',
-      //   headerTitleAlign: 'center',
-      //   animation: 'fade_from_bottom',
-      //   headerRight: () => (
-      //     (user?.uid === route.params.thread.initialUser) &&
-      //     <Icon name="trash" size={24} color="#fff"
-      //       onPress={() => setShowAuthModal(route.params.thread)}
-      //     />
-      //   )
-      // })}
-      />
-    </Drawer.Navigator>
-  )
 }
