+++ jin/screens/Chatroom.js
@@ -6,21 +6,13 @@ import {
   FlatList,
 } from 'react-native';
 import Text from '../../defaultSetting/FontText';
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { NativeBaseProvider, Stack, HStack, Modal, Button, useToast, Spinner } from 'native-base';
-import Icon from 'react-native-vector-icons/FontAwesome';
-import auth from '@react-native-firebase/auth';
+import { NativeBaseProvider, Stack, HStack, useToast, Spinner } from 'native-base';
 import firestore from '@react-native-firebase/firestore';
 
-import Authentication from './Authentication';
-import CreateChat from "./CreateChat";
-import Chat from "./Chat";
 // import HeaderClassicSearchBar from "../../lib/src/HeaderClassicSearchBar/HeaderClassicSearchBar";
 const [headerColor, iconActiveColor, iconInActiveColor] = ["#efefef", "#BF2A52", "#bbb"];
 
-const StackNav = createNativeStackNavigator();
-
-function Chatroom({ navigation, route }) {
+export default function Chatroom({ navigation, route }) {
   const [threads, setThreads] = useState([]);
   const [loading, setLoading] = useState(true);
   const [timer, setTimer] = useState(new Date().getMinutes());
@@ -163,96 +155,6 @@ function Chatroom({ navigation, route }) {
   );
 }
 
-export default function ({ navigation }) {
-  const user = auth().currentUser;
-  const [showAuthModal, setShowAuthModal] = useState(false);
-
-  function deleteChat(id) {
-    firestore()
-      .collection('Chat')
-      .doc(id)
-      .delete().then(() => {
-        setShowAuthModal(false);
-        navigation.navigate('같이 배달 리스트', { response: 0 });
-      })
-  }
-
-  return (
-    <NativeBaseProvider>
-      <StackNav.Navigator>
-        <StackNav.Screen
-          name="같이 배달 리스트"
-          component={Chatroom}
-          initialParams={{ response: -1 }}
-          options={{
-            headerTitle: () => (
-              <Text style={styles.headerTitle} bold>같이 배달</Text>
-            ),
-            headerStyle: {
-              backgroundColor: '#BF2A52',
-            },
-            headerTintColor: '#fff',
-            headerTitleAlign: 'center',
-            headerRight: () => (
-              <Icon name="plus" size={24} color="#fff"
-                onPress={() => navigation.navigate('새로운 채팅방 만들기')} />
-            )
-          }} />
-        <StackNav.Screen
-          name="새로운 채팅방 만들기"
-          component={CreateChat}
-          options={{
-            headerTitle: () => (
-              <Text style={styles.headerTitle} bold>{"새로운 채팅방 만들기"}</Text>
-            ),
-            headerBackTitleVisible: false,
-            headerStyle: {
-              backgroundColor: '#BF2A52',
-            },
-            headerTintColor: '#fff',
-            headerTitleAlign: 'center',
-            animation: 'slide_from_right'
-          }} />
-        <StackNav.Screen
-          name="메시지"
-          component={Chat}
-          options={({ navigation, route }) => ({
-            headerTitle: () => (
-              <Text style={styles.headerTitle} bold>{route.params.thread.name}</Text>
-            ),
-            headerBackTitleVisible: false,
-            headerStyle: {
-              backgroundColor: '#BF2A52',
-            },
-            headerTintColor: '#fff',
-            headerTitleAlign: 'center',
-            animation: 'fade_from_bottom',
-            headerRight: () => (
-              (user?.uid === route.params.thread.initialUser) &&
-              <Icon name="trash" size={24} color="#fff"
-                // onPress={() => setShowAuthModal(route.params.thread)}
-                onPress={() => navigation.dispatch(DrawerActions.toggleDrawer())}
-              />
-            )
-          })}
-        />
-      </StackNav.Navigator>
-      <Modal isOpen={!!showAuthModal} onClose={() => setShowAuthModal(false)}>
-        <Modal.Content style={styles.modalContent}>
-          <Modal.Header>정말로 채팅방을 삭제하시겠습니까?</Modal.Header>
-          <Modal.Body>채팅방을 지우면 다시 되돌릴 수 없습니다. 신중히 선택해주세요.</Modal.Body>
-          <Modal.Footer>
-            <Button.Group variant="ghost">
-              <Button onPress={() => deleteChat(showAuthModal._id)}>예</Button>
-              <Button onPress={() => setShowAuthModal(false)}>아니요</Button>
-            </Button.Group>
-          </Modal.Footer>
-        </Modal.Content>
-      </Modal>
-    </NativeBaseProvider>
-  )
-}
-
 const styles = StyleSheet.create({
   container: {
     flex: 1,
@@ -313,18 +215,6 @@ const styles = StyleSheet.create({
     fontSize: 16,
     marginVertical: 6,
     marginLeft: 5
-  },
-  modalContent: {
-    minWidth: 100,
-    maxWidth: 250,
-    minHeight: 100,
-    maxHeight: 250,
-    justifyContent: 'center',
-    alignItems: 'center'
-  },
-  headerTitle: {
-    fontSize: 20,
-    color: '#f5f5f5'
   },
   emptyView: {
     flexDirection: 'row',
